requirements.txt
asgiref==3.8.1
mysql-connector-python==8.3.0
Django==5.0.4
djangorestframework==3.15.1
django-filter==24.2
pillow==10.3.0
sqlparse==0.4.4
tzdata==2024.1
